CUDA-C-PROTOTYPE: Added splitting of generator powers based on MVAbase only and not adhering to Pg and Qg limits.
The previous code, although it is the correct one and should be always used, gave a non-zero reactive power mismatch at some buses that had multiple generators. The issue here is that the total reactive power generation required at the bus exceeded the total VAR limit of the generators and hence the non-zero residual. As an alternate, for those who do not to stick with VAR limits (for whatever reasons), the split_gen_powe_within_limits option is added to PFLOW. When it is true, it will use the proportional splitting of generator (real and) reactive power. If not, it will simply split the power based on the MVAbase.