Update OPFLOW IPOPT solver implementation
The OPFLOW-IPOPT solver implementatin creates a symmetric matrix (SeqSBAIJ) that
is the format needed for the IPOPT solver. Earlier this was done explicitly done
by the formulations. Refactored the Hessian code for all formulations to calculate
the entire Hessian piece for the equality and inequality constraints. If the solver,
such as IPOPT, needs a symmetric sparse format then it should be doing the conversion,
not the formulation. This change will allow other solvers, like Tao, that do not use
a symmetric sparse format to directly work with the formulations.