SCOPFLOW: Fixed valgrind errors with PIPS.
PIPS requires what it calls "parent contributions" for the Hessian. These
are the contributions from the second-stage variables to the Hessian for the
first stage. These HAVE(?) to be set even if there are no contributions from
the second-stage variables. In this case, all zeros need to be set in the
Hessian buffer (elts) supplied by PIPS.