SCOPFLOW: Added 'model' to SCOPFLOW. Only used by solver IPOPTNEW.
IPOPTNEW is the revised version of SCOPFLOW IPOPT solver. It calls the model for doing
the model evaluation - similar to what is done for OPFLOW.

For large matrices, this becomes very slow compared to IPOPT only (observed this for
the 118 bus system with all contingencies) because the matrix setup takes a very long time since
there is no preallocation done. One possibility to improve this is to do preallocation. But, this
will have to be done for OPFLOW as well.