SOPFLOW bug-fix: Scenarios producing same cost. Resolves !186
The bounds were not getting enforced because they were changed on reading the scenarios
which was done AFTER the optimization solver was created. Optimization solvers DO NOT ALLOW
changing bounds (variables and constraints) after the solver object is created. Updated the
sopflow code to read the scenarios before the solver object is cretaed (in OPFLOWSetUp). Also,
fixed a bug that was causing incorrect reading of the last scenario.