Fix ignoring contingencies when the from bus number is greater than than to bus number (#70)
* Fix ignoring contingencies when the from bus number is greater than to bus #26

When the from bus number is greater than the to bus number for a branch, the
PS branch object stores it in the reverse order. This is done so that the
from bus number is always less than to bus number. The reason for doing this
is IPOPT complains for Hessian if this is not done. However, this was causing
the contingencies with from bus > to bus to be ignored.

* Apply pre-commmit fixes

* Update CMakeLists.txt

---------

Co-authored-by: abhyshr <abhyshr@users.noreply.github.com>
Co-authored-by: Cameron Rutherford <robert.rutherford@pnnl.gov>